Rewrite check() with a pipelined implementation
This will enable:
 - verification and correction of sizes and offsets;
 - parallel decompression of bgzf blocks payloads.

At this point, support for gzip extra subfields has been lost and performance
is around 5 ~ 10% better than with the previous implementation. Computation of
byte counts has been disabled too, but I'm not yet decided if I'll keep it
anyway, given how unreliable it is to estimate the number of inflated corrupted
bytes.