Handle deflate and CRC32 in another thread, synchronously
Obviously, this needs to be done asynchronously, since this version is
about twice slower. Attention must be paid to wait for the futures in streaming
and not at the end of the bam file, not only for performance reasons, but also
because it could cost a lot of memory on huge bam files.