Handle deflate and CRC32 on a user-specified number of threads
The number of threads is either specified using the new --threads flag or
automatically set to the number of actually available hardware threads on the
machine.

Up to 100 threads are allowed to work in parallel. The goal is more to let each
thread finish its work before being waited for than to gain any extra speedup
with more parallelism, give how small the gzip blocks are (at most 64 kiB).
100 blocks costs very little in RAM (less than 10 MiB at worst) and more threads
wouldn't help anyway, so this seems like a fair number -- with a large margin.

This version is actually faster than gzip on an Intel Core i5 6500 with two
threads, and *much* faster with four threads, even if the implementation has
not yet been optimized at all.

This closes #1