Handle limited corruption in gzip / bgzf headers
Accept up to 1/4 + 1/6 corrupted bytes in some conditions. This allows to
rescue bam files with single-byte corruptions in almost any location.

Since the inflated payload always ends up being checked against its CRC32,
there is no risk to miss an impacting corruption there; it just improves on
the chances to rescue a bgzf block the header of which has been corrupted.

Corruptions in the gzip extra field length field, in the bgzf payload length
field and all kinds of corruptions in the gzip / bgzf header that impact more
bytes still yield a panic!() for the time being.

This represents some progress regarding #2