Handle remaining cases of corruption in gzip / bgzf headers
When the gzip header is not here (or broken enough not to be found), look for
a valid gzip header anywhere after the start of the previous block.

When the gzip header is here but broken, look for a valid gzip header anywhere
after the start of the current block.

Stop counting bad blocks twice when some blocks have multiple corruptions.

This closes #2